+++ packages/testing/playwright/currents.config.ts
@@ -2,7 +2,7 @@ import type { CurrentsConfig } from '@currents/playwright';
 
 const config: CurrentsConfig = {
 	recordKey: process.env.CURRENTS_RECORD_KEY ?? '',
-	projectId: process.env.CURRENTS_PROJECT_ID ?? 'qyhJh8',
+	projectId: process.env.CURRENTS_PROJECT_ID ?? 'I0yzoc',
 	...(process.env.BUILD_WITH_COVERAGE === 'true' && {
 		coverage: {
 			projects: true,
