+++ packages/testing/playwright/playwright-projects.ts
@@ -25,7 +25,10 @@ const CONTAINER_CONFIGS: Array<{ name: string; config: N8NConfig }> = [
 	{ name: 'sqlite', config: {} },
 	{ name: 'postgres', config: { postgres: true } },
 	{ name: 'queue', config: { workers: 1 } },
-	{ name: 'multi-main', config: { mains: 2, workers: 1, services: ['observability'] } },
+	{
+		name: 'multi-main',
+		config: { mains: 2, workers: 1, services: ['victoriaLogs', 'victoriaMetrics', 'vector'] },
+	},
 ];
 
 export function getProjects(): Project[] {
