+++ packages/testing/playwright/tests/performance/memory-consumption-cloud.spec.ts
@@ -7,7 +7,7 @@ test.use({
 			memory: 0.75,
 			cpu: 0.5,
 		},
-		services: ['observability'],
+		services: ['victoriaLogs', 'victoriaMetrics', 'vector'],
 	},
 });
